Backend: Rename couchers.i18n.i18n to couchers.i18n.localize
