+++ app/backend/src/couchers/notifications/render_push.py
@@ -6,7 +6,7 @@ import logging
 from typing import Any, assert_never
 
 from couchers import urls
-from couchers.i18n.i18n import format_phone_number, localize_date_from_iso, localize_datetime_for_user
+from couchers.i18n.localize import format_phone_number, localize_date_from_iso, localize_datetime_for_user
 from couchers.models import Notification, NotificationTopicAction, User
 from couchers.notifications.push import PushNotificationContent
 from couchers.proto import events_pb2, notification_data_pb2
